planner_suggest: change according to review
